refactor:TOP-58 data/remote/service 패키지 정리
