+++ data/src/main/java/com/tht/tht/data/remote/service/image/ImageServiceImpl.kt
@@ -1,4 +1,4 @@
-package com.tht.tht.data.remote.service
+package com.tht.tht.data.remote.service.image
 
 import android.net.Uri
 import android.util.Log
