+++ data/src/test/java/com/tht/tht/data/datasource/RegionCodeDataSourceImplTest.kt
@@ -5,7 +5,7 @@ import com.tht.tht.data.remote.mapper.toUnwrap
 import com.tht.tht.data.remote.response.base.BaseResponse
 import com.tht.tht.data.remote.response.base.ErrorResponse
 import com.tht.tht.data.remote.response.regioncode.RegionCodeResponse
-import com.tht.tht.data.remote.service.RegionCodeApi
+import com.tht.tht.data.remote.service.location.RegionCodeApi
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -13,8 +13,6 @@ import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestCoroutineScheduler
 import kotlinx.coroutines.test.runTest
 import org.assertj.core.api.Assertions.assertThat
-import org.assertj.core.api.Assertions.assertThatThrownBy
-import org.junit.Assert.assertThrows
 import org.junit.Before
 import org.junit.Test
